+++ tests/e2e/framework/kubernetes.go
@@ -15,6 +15,7 @@
 package framework
 
 import (
+	"context"
 	"flag"
 	"fmt"
 	"io/ioutil"
@@ -29,12 +30,13 @@ import (
 	"testing"
 	"time"
 
-	"github.com/hashicorp/go-multierror"
+	multierror "github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 
-	"istio.io/common/pkg/log"
 	testKube "istio.io/istio/pkg/test/kube"
+	"istio.io/istio/pkg/test/util/retry"
 	"istio.io/istio/tests/util"
+	"istio.io/pkg/log"
 )
 
 const (
@@ -53,6 +55,7 @@ const (
 	mcAuthInstallFileNamespace     = "istio-auth-multicluster.yaml"
 	mcRemoteInstallFile            = "istio-remote.yaml"
 	mcSplitHorizonInstallFile      = "istio-multicluster-split-horizon.yaml"
+	authOperatorInstallFile        = "istio-operator.yaml"
 	istioSystem                    = "istio-system"
 	istioIngressServiceName        = "istio-ingress"
 	istioIngressLabel              = "ingress"
@@ -62,7 +65,7 @@ const (
 	defaultSidecarInjectorFile     = "istio-sidecar-injector.yaml"
 	ingressCertsName               = "istio-ingress-certs"
 	maxDeploymentRolloutTime       = 960 * time.Second
-	maxValidationReadyCheckTime    = 30 * time.Second
+	maxValidationReadyCheckTime    = 120 * time.Second
 	maxCNIDeployTime               = 10 * time.Second
 	helmServiceAccountFile         = "helm-service-account.yaml"
 	istioHelmInstallDir            = istioInstallDir + "/helm/istio"
@@ -76,16 +79,19 @@ const (
 	// CRD files that should be installed during testing
 	// NB: these files come from the directory install/kubernetes/helm/istio-init/files/*crd*
 	//     and contain all CRDs used by Istio during runtime
-	zeroCRDInstallFile        = "crd-10.yaml"
-	oneCRDInstallFile         = "crd-11.yaml"
-	twoCRDInstallFile         = "crd-12.yaml"
+	allCRDInstallFile         = "crd-all.gen.yaml"
+	mixerCRDInstallFile       = "crd-mixer.yaml"
 	certManagerCRDInstallFile = "crd-certmanager-10.yaml"
 	// PrimaryCluster identifies the primary cluster
 	PrimaryCluster = "primary"
 	// RemoteCluster identifies the remote cluster
 	RemoteCluster = "remote"
 
+	kubernetesReadinessTimeout        = time.Second * 180
+	kubernetesReadinessInterval       = 200 * time.Millisecond
 	validationWebhookReadinessTimeout = time.Minute
+	istioOperatorTimeout              = time.Second * 300
+	istioOperatorFreq                 = time.Second * 10
 	validationWebhookReadinessFreq    = 100 * time.Millisecond
 )
 
@@ -96,6 +102,8 @@ var (
 	mixerTag           = flag.String("mixer_tag", os.Getenv("TAG"), "Mixer tag")
 	pilotHub           = flag.String("pilot_hub", os.Getenv("HUB"), "Pilot hub")
 	pilotTag           = flag.String("pilot_tag", os.Getenv("TAG"), "Pilot tag")
+	appHub             = flag.String("app_hub", os.Getenv("HUB"), "Test application hub")
+	appTag             = flag.String("app_tag", os.Getenv("TAG"), "Test application tag")
 	proxyHub           = flag.String("proxy_hub", os.Getenv("HUB"), "Proxy hub")
 	proxyTag           = flag.String("proxy_tag", os.Getenv("TAG"), "Proxy tag")
 	caHub              = flag.String("ca_hub", os.Getenv("HUB"), "Ca hub")
@@ -122,6 +130,7 @@ var (
 	useGalleyConfigValidator = flag.Bool("use_galley_config_validator", true, "Use galley configuration validation webhook")
 	installer                = flag.String("installer", "kubectl", "Istio installer, default to kubectl, or helm")
 	useMCP                   = flag.Bool("use_mcp", true, "use MCP for configuring Istio components")
+	useOperator              = flag.Bool("use_operator", false, "use Operator to deploy Istio components")
 	outboundTrafficPolicy    = flag.String("outbound_trafficpolicy", "ALLOW_ANY", "Istio outbound traffic policy, default to ALLOW_ANY")
 	enableEgressGateway      = flag.Bool("enable_egressgateway", false, "enable egress gateway, default to false")
 	useCNI                   = flag.Bool("use_cni", false,
@@ -206,6 +215,8 @@ func getClusterWideInstallFile() string {
 		if *useMCP {
 			if *authSdsEnable {
 				istioYaml = authSdsInstallFile
+			} else if *useOperator {
+				istioYaml = authOperatorInstallFile
 			} else {
 				istioYaml = authInstallFile
 			}
@@ -370,11 +381,15 @@ func (k *KubeInfo) IstioEgressGatewayService() string {
 	return istioEgressGatewayServiceName
 }
 
-// Setup set up Kubernetes prerequest for tests
+// Setup Kubernetes pre-requisites for tests
 func (k *KubeInfo) Setup() error {
 	log.Infoa("Setting up kubeInfo setupSkip=", *skipSetup)
 	var err error
 	if err = os.Mkdir(k.yamlDir, os.ModeDir|os.ModePerm); err != nil {
+		return err
+	}
+
+	if err = k.waitForKubernetes(); err != nil {
 		return err
 	}
 
@@ -423,6 +438,16 @@ func (k *KubeInfo) PilotHub() string {
 // PilotTag exposes the Docker tag used for the pilot image.
 func (k *KubeInfo) PilotTag() string {
 	return *pilotTag
+}
+
+// AppHub exposes the Docker hub used for the test application image.
+func (k *KubeInfo) AppHub() string {
+	return *appHub
+}
+
+// AppTag exposes the Docker tag used for the test application image.
+func (k *KubeInfo) AppTag() string {
+	return *appTag
 }
 
 // ProxyHub exposes the Docker hub used for the proxy image.
@@ -531,8 +556,8 @@ func (k *KubeInfo) Teardown() error {
 			}
 		}
 
+		var istioYaml string
 		if *clusterWide {
-			var istioYaml string
 			if *multiClusterDir != "" {
 				if *authEnable {
 					istioYaml = mcAuthInstallFileNamespace
@@ -542,13 +567,34 @@ func (k *KubeInfo) Teardown() error {
 			} else {
 				istioYaml = getClusterWideInstallFile()
 			}
-
+		}
+		if *useOperator {
+			//save operator logs
+			log.Info("Saving istio-operator logs")
+			if err := util.FetchAndSaveClusterLogs("istio-operator", k.TmpDir, k.KubeConfig); err != nil {
+				log.Errorf("Failed to save operator logs: %v", err)
+			}
+			// Need an operator unique delete procedure
+			if _, err := util.Shell("kubectl -n istio-operator delete IstioOperator example-istiocontrolplane"); err != nil {
+				log.Errorf("Failed to delete the Istio CR.")
+				return err
+			}
+			if _, err := util.Shell("kubectl delete ns istio-operator --kubeconfig=%s",
+				k.KubeConfig); err != nil {
+				log.Errorf("Failed to delete istio-operator namespace.")
+				return err
+			}
+			if _, err := util.Shell("kubectl delete ns %s --kubeconfig=%s",
+				k.Namespace, k.KubeConfig); err != nil {
+				log.Errorf("Failed to delete %s namespace.", k.Namespace)
+				return err
+			}
+		} else {
 			testIstioYaml := filepath.Join(k.TmpDir, "yaml", istioYaml)
-
 			if err := util.KubeDelete(k.Namespace, testIstioYaml, k.KubeConfig); err != nil {
 				log.Infof("Safe to ignore resource not found errors in kubectl delete -f %s", testIstioYaml)
 			}
-		} else {
+
 			if err := util.DeleteNamespace(k.Namespace, k.KubeConfig); err != nil {
 				log.Errorf("Failed to delete namespace %s", k.Namespace)
 				return err
@@ -585,7 +631,11 @@ func (k *KubeInfo) Teardown() error {
 	validatingWebhookConfigurationExists := false
 	log.Infof("Deleting namespace %v", k.Namespace)
 	for attempts := 1; attempts <= maxAttempts; attempts++ {
-		namespaceDeleted, _ = util.NamespaceDeleted(k.Namespace, k.KubeConfig)
+		if *useOperator {
+			namespaceDeleted, _ = util.NamespaceDeleted("istio-operator", k.KubeConfig)
+		} else {
+			namespaceDeleted, _ = util.NamespaceDeleted(k.Namespace, k.KubeConfig)
+		}
 		// As validatingWebhookConfiguration "istio-galley" will
 		// be delete by kubernetes GC controller asynchronously,
 		// we need to ensure it's deleted before return.
@@ -630,6 +680,33 @@ func (k *KubeInfo) GetAppPods(cluster string) map[string][]string {
 		}
 	}
 	return newMap
+}
+
+// CheckJobSucceeded checks whether the job succeeded.
+func (k *KubeInfo) CheckJobSucceeded(cluster, jobName string) error {
+	retry := util.Retrier{
+		BaseDelay: 1 * time.Second,
+		MaxDelay:  1 * time.Second,
+		Retries:   15,
+	}
+
+	retryFn := func(_ context.Context, i int) error {
+		ret, err := util.IsJobSucceeded(k.Namespace, jobName, k.Clusters[cluster])
+		if err != nil {
+			log.Errorf("Failed to get retrieve the app pods for namespace %s", k.Namespace)
+			return err
+		}
+		if !ret {
+			return fmt.Errorf("job %s not succeeded", jobName)
+		}
+		return nil
+	}
+	ctx := context.Background()
+	_, err := retry.Retry(ctx, retryFn)
+	if err != nil {
+		return err
+	}
+	return nil
 }
 
 // GetRoutes gets routes from the pod or returns error
@@ -707,12 +784,6 @@ func (k *KubeInfo) deployIstio() error {
 		}
 	}
 
-	// Create istio-system namespace
-	if err := util.CreateNamespace(k.Namespace, k.KubeConfig); err != nil {
-		log.Errorf("Unable to create namespace %s: %s", k.Namespace, err.Error())
-		return err
-	}
-
 	// Deploy the CNI if enabled
 	if *useCNI {
 		err := k.deployCNI()
@@ -743,41 +814,66 @@ func (k *KubeInfo) deployIstio() error {
 
 	}
 
-	// Apply istio-init
-	yamlDir := filepath.Join(istioInstallDir, initInstallFile)
-	baseIstioYaml := filepath.Join(k.ReleaseDir, yamlDir)
-	testIstioYaml := filepath.Join(k.TmpDir, "yaml", istioYaml)
-	if err := k.generateIstio(baseIstioYaml, testIstioYaml); err != nil {
-		log.Errorf("Generating istio-init.yaml")
-		return err
-	}
-
-	if err := util.KubeApply(k.Namespace, testIstioYaml, k.KubeConfig); err != nil {
-		log.Errorf("istio-init.yaml  %s deployment failed", testIstioYaml)
-		return err
-	}
-
-	// TODO(sdake): need a better synchronization
-	time.Sleep(20 * time.Second)
-
-	// Apply main manifest
-	yamlDir = filepath.Join(istioInstallDir, istioYaml)
-	baseIstioYaml = filepath.Join(k.ReleaseDir, yamlDir)
-	testIstioYaml = filepath.Join(k.TmpDir, "yaml", istioYaml)
-
-	if err := k.generateIstio(baseIstioYaml, testIstioYaml); err != nil {
-		log.Errorf("Generating yaml %s failed", testIstioYaml)
-		return err
-	}
-
-	if *multiClusterDir != "" {
-		if err := k.createCacerts(false); err != nil {
-			log.Infof("Failed to create Cacerts with namespace %s in primary cluster", k.Namespace)
+	var testIstioYaml string
+	// Use the operator manifest when operator mode enabled
+	if *useOperator {
+		istioYaml = authOperatorInstallFile
+		yamlDir := filepath.Join(istioInstallDir, istioYaml)
+		baseIstioYaml := filepath.Join(k.ReleaseDir, yamlDir)
+		testIstioYaml = filepath.Join(k.TmpDir, "yaml", istioYaml)
+		util.CopyFile(baseIstioYaml, testIstioYaml)
+		if err := util.KubeApply("istio-operator", testIstioYaml, k.KubeConfig); err != nil {
+			log.Errorf("Istio operator %s deployment failed", testIstioYaml)
+			return err
 		}
-	}
-	if err := util.KubeApply(k.Namespace, testIstioYaml, k.KubeConfig); err != nil {
-		log.Errorf("Istio core %s deployment failed", testIstioYaml)
-		return err
+		if err := k.waitForIstioOperator(); err != nil {
+			log.Errorf("istio operator fails to deploy Istio: %v", err)
+			return err
+		}
+	} else {
+		// Create istio-system namespace
+		if err := util.CreateNamespace(k.Namespace, k.KubeConfig); err != nil {
+			log.Errorf("Unable to create namespace %s: %s", k.Namespace, err.Error())
+			return err
+		}
+
+		// Apply istio-init and generate an Istio manifest for later
+		yamlDir := filepath.Join(istioInstallDir, initInstallFile)
+		baseIstioYaml := filepath.Join(k.ReleaseDir, yamlDir)
+		testIstioYaml = filepath.Join(k.TmpDir, "yaml", istioYaml)
+		if err := k.generateIstio(baseIstioYaml, testIstioYaml); err != nil {
+			log.Errorf("Generating istio-init.yaml")
+			return err
+		}
+
+		if err := util.KubeApply(k.Namespace, testIstioYaml, k.KubeConfig); err != nil {
+			log.Errorf("istio-init.yaml  %s deployment failed", testIstioYaml)
+			return err
+		}
+
+		// TODO(sdake): need a better synchronization
+		time.Sleep(20 * time.Second)
+
+		// Generate main manifest for application in the rest of this function
+		// depending on E2E flags
+		yamlDir = filepath.Join(istioInstallDir, istioYaml)
+		baseIstioYaml = filepath.Join(k.ReleaseDir, yamlDir)
+		testIstioYaml = filepath.Join(k.TmpDir, "yaml", istioYaml)
+
+		if err := k.generateIstio(baseIstioYaml, testIstioYaml); err != nil {
+			log.Errorf("Generating yaml %s failed", testIstioYaml)
+			return err
+		}
+
+		if *multiClusterDir != "" {
+			if err := k.createCacerts(false); err != nil {
+				log.Infof("Failed to create Cacerts with namespace %s in primary cluster", k.Namespace)
+			}
+		}
+		if err := util.KubeApply(k.Namespace, testIstioYaml, k.KubeConfig); err != nil {
+			log.Errorf("Istio core %s deployment failed", testIstioYaml)
+			return err
+		}
 	}
 
 	if *multiClusterDir != "" {
@@ -898,6 +994,17 @@ spec:
 `
 )
 
+// Wait for Kubernetes to become active within kubernetesReadinessTimeout period
+// This operation only retreives the pods in the kube-system namespace
+// (TODO) sdake: This may be insufficient for a complete readiness check of Kubernetes
+func (k *KubeInfo) waitForKubernetes() error {
+	log.Info("Waiting for Kubernetes to become responsive")
+	return retry.UntilSuccess(func() error {
+		_, err := k.KubeAccessor.GetPods("kube-system")
+		return err
+	}, retry.Delay(kubernetesReadinessInterval), retry.Timeout(kubernetesReadinessTimeout))
+}
+
 func (k *KubeInfo) waitForValdiationWebhook() error {
 
 	add := fmt.Sprintf(`cat << EOF | kubectl --kubeconfig=%s apply -f -
@@ -929,6 +1036,27 @@ EOF`, k.KubeConfig, dummyValidationRule)
 	return nil
 }
 
+func (k *KubeInfo) waitForIstioOperator() error {
+
+	get := fmt.Sprintf(`kubectl --kubeconfig=%s get iop example-istiocontrolplane -n istio-operator -o yaml`, k.KubeConfig)
+	timeout := time.Now().Add(istioOperatorTimeout)
+	for {
+		if time.Now().After(timeout) {
+			return errors.New("timeout waiting for istio operator to deploy Istio")
+		}
+		out, err := util.ShellSilent(get)
+		if err == nil && strings.Contains(out, "HEALTHY") {
+			break
+		}
+
+		log.Warnf("istio-operator is still deploying Istio: %v", err)
+		time.Sleep(istioOperatorFreq)
+
+	}
+	log.Info("istio operator succeeds to deploy Istio")
+	return nil
+}
+
 func (k *KubeInfo) deployCRDs(kubernetesCRD string) error {
 	yamlFileName := filepath.Join(istioInstallDir, helmInstallerName, "istio-init", "files", kubernetesCRD)
 	yamlFileName = filepath.Join(k.ReleaseDir, yamlFileName)
@@ -943,7 +1071,7 @@ func (k *KubeInfo) deployCRDs(kubernetesCRD string) error {
 func (k *KubeInfo) deployIstioWithHelm() error {
 	// Note: When adding a CRD to the install, a new CRDFile* constant is needed
 	// This slice contains the list of CRD files installed during testing
-	istioCRDFileNames := []string{zeroCRDInstallFile, oneCRDInstallFile, twoCRDInstallFile, certManagerCRDInstallFile}
+	istioCRDFileNames := []string{allCRDInstallFile, mixerCRDInstallFile, certManagerCRDInstallFile}
 	// deploy all CRDs in Istio first
 	for _, yamlFileName := range istioCRDFileNames {
 		if err := k.deployCRDs(yamlFileName); err != nil {
@@ -966,10 +1094,12 @@ func (k *KubeInfo) deployIstioWithHelm() error {
 		setValue += " --set sidecarInjectorWebhook.enabled=true"
 	}
 
-	if *useMCP {
-		setValue += " --set galley.enabled=true --set global.useMCP=true"
+	if *useMCP && *useGalleyConfigValidator {
+		setValue += " --set galley.enabled=true --set global.useMCP=true --set global.configValidation=true"
+	} else if *useMCP {
+		setValue += " --set galley.enabled=true --set global.useMCP=true --set global.configValidation=false"
 	} else if *useGalleyConfigValidator {
-		setValue += " --set galley.enabled=true --set global.useMCP=false"
+		setValue += " --set galley.enabled=true --set global.useMCP=false --set global.configValidation=true"
 	} else {
 		setValue += " --set galley.enabled=false --set global.useMCP=false"
 	}
@@ -985,7 +1115,7 @@ func (k *KubeInfo) deployIstioWithHelm() error {
 	// hubs and tags replacement.
 	// Helm chart assumes hub and tag are the same among multiple istio components.
 	if *pilotHub != "" && *pilotTag != "" {
-		setValue = setValue + " --set global.hub=" + *pilotHub + " --set global.tag=" + *pilotTag
+		setValue += " --set-string global.hub=" + *pilotHub + " --set-string global.tag=" + *pilotTag
 	}
 
 	if !*clusterWide {
@@ -1085,7 +1215,6 @@ func (k *KubeInfo) generateSidecarInjector(src, dst string) error {
 	if *pilotHub != "" && *pilotTag != "" {
 		content = updateImage("sidecar_injector", *pilotHub, *pilotTag, content)
 		content = updateInjectVersion(*pilotTag, content)
-		content = updateInjectImage("initImage", "proxy_init", *proxyHub, *proxyTag, content)
 		content = updateInjectImage("proxyImage", "proxy", *proxyHub, *proxyTag, content)
 	}
 
@@ -1103,6 +1232,10 @@ func replacePattern(content []byte, src, dest string) []byte {
 	return content
 }
 
+// This code is in need of a reimplementation and some rethinking. An in-place
+// modification on the raw manifest is the wrong approach. This model is also
+// not portable to our future around IstioOperator where we don't necessarily
+// have a manifest to in place modify...
 func (k *KubeInfo) generateIstio(src, dst string) error {
 	content, err := ioutil.ReadFile(src)
 	if err != nil {
@@ -1136,10 +1269,6 @@ func (k *KubeInfo) generateIstio(src, dst string) error {
 		}
 		if *pilotHub != "" && *pilotTag != "" {
 			content = updateImage("pilot", *pilotHub, *pilotTag, content)
-		}
-		if *proxyHub != "" && *proxyTag != "" {
-			//Need to be updated when the string "proxy" is changed as the default image name
-			content = updateImage("proxy_init", *proxyHub, *proxyTag, content)
 		}
 		if *proxyHub != "" && *proxyTag != "" {
 			//Need to be updated when the string "proxy" is changed as the default image name
@@ -1205,7 +1334,7 @@ func (k *KubeInfo) deployCNI() error {
 	log.Info("Deploy Istio CNI components")
 	// Some environments will require additional options to be set or changed
 	// (e.g. GKE environments need the bin directory to be changed from the default
-	setValue := " --set hub=" + *cniHub + " --set tag=" + *cniTag
+	setValue := " --set-string hub=" + *cniHub + " --set-string tag=" + *cniTag
 	setValue += " --set excludeNamespaces={} --set pullPolicy=IfNotPresent --set logLevel=debug"
 	if extraHelmValues := os.Getenv("EXTRA_HELM_SETTINGS"); extraHelmValues != "" {
 		setValue += extraHelmValues
